+++ packages/orq-rc/RELEASES.md
@@ -7399,3 +7399,13 @@ Based on:
 - [typescript v4.5.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.19 - packages/orq-rc
+
+## 2026-03-02 14:27:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.736.3 (2.845.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.21 - packages/orq-rc